chore(phase3): coordinator pre-work — utils, getJobsListData cap, perf indexes
Lands the three Luke/Farhan blockers that gate the Phase 3 build agents,
ahead of dispatching them.

- web/lib/scrapers/utils.ts: extracted `decodeHtmlEntities` as the single
  canonical implementation. The two private copies in phenom-rbc.ts and
  scotiabank.ts are gone; phenom-rbc re-exports for backward compat.
  Scotiabank's copy had `&amp;` decoded FIRST — a real bug that caused
  doubly-encoded sequences like `&amp;lt;` to collapse to `<` instead of
  staying as `&lt;`. The canonical version decodes `&amp;` last.
- web/lib/dashboard-queries.ts: `getJobsListData` now hard-caps at
  JOBS_LIST_MAX_ROWS = 500 most-recent. Without it the /jobs page would
  ship ~22 MB of JSON on every `?tier=incumbent` or `?tier=all` render
  once the 5-bank ramp finishes. Server-side pagination is the proper
  follow-up.
- web/supabase/migrations/20260522000000_phase3_indexes.sql: two partial
  indexes — `idx_job_postings_company_active_first_seen` (covers
  getIncumbentBets + getCompanyHiringSparkline/Delta/Pivot) and
  `idx_companies_tier_active` (every cross-company aggregator's tier
  join). Migration is not yet applied — apply before the bank corpus
  ramps to keep query plans clean.

37 tests pass; build clean.
